[misc] Meshcat is now an optional requirement. (#512)
* [misc] Check build type is supported in deps install scripts.
* [misc] Meshcat is now an optional requirement.
* [misc] Matplotlib is now an optional requirement.
