+++ docs/api/jiminy_py/log.rst
@@ -5,4 +5,3 @@ log
    :members:
    :undoc-members:
    :show-inheritance:
-   :exclude-members: plot_log
