+++ docs/api/jiminy_py/plot.rst
@@ -5,4 +5,4 @@ plot
    :members:
    :undoc-members:
    :show-inheritance:
-
+   :exclude-members: plot_log
