+++ python/jiminy_py/setup.py
@@ -26,6 +26,8 @@ if np_ver < (1, 20):
     np_req = "numpy<1.20"
 else:
     np_req = "numpy>=1.20,!=1.21.0,!=1.21.1,!=1.21.2,!=1.21.3,!=1.21.4"
+    if np_ver < (1, 22):
+        np_req += ",<1.22"
 
 
 setup(
@@ -75,14 +77,12 @@ setup(
     ],
     include_package_data=True,
     entry_points={"console_scripts": [
-        "jiminy_plot=jiminy_py.log:plot_log",
+        "jiminy_plot=jiminy_py.plot:plot_log",
         ("jiminy_meshcat_server="
          "jiminy_py.meshcat.server:start_meshcat_server_standalone"),
         "jiminy_replay=jiminy_py.viewer.replay:_play_logs_files_entrypoint"
     ]},
     install_requires=[
-        # Used internally by Viewer to read/write snapshots.
-        "pillow",
         # Add support of TypedDict to any Python 3 version.
         # 3.10.0 adds 'ParamSpec' that is required for pylint>=2.11.1.
         "typing_extensions>=3.10.0",
@@ -90,42 +90,21 @@ setup(
         "tqdm",
         # Standard library for matrix algebra.
         # >=1.20 breaks ABI
-        # >=1.21,<1.21.5 is causing segfault with boost::python
+        # >=1.21,<1.21.5 is causing segfault with boost::python.
+        #     See issue: https://github.com/boostorg/python/issues/376
+        # >= 1.22 breaks API for compiled libs.
         np_req,
-        # Use to operate on nested data structure conveniently
-        # - 0.1.5 introduces `tree.traverse` method that it used to operate on
-        # `gym.spaces.Dict`.
-        # - 0.1.6 adds support of Python 3.9 and unifies API method naming.
-        "dm-tree>=0.1.6",
-        # Used internally for interpolation and filtering.
-        # No wheel is distributed for PyPy on pypi, and pip is only able to
-        # build from source after install `libatlas-base-dev` system
-        # dependency.
-        # 1.2.0 fixes `fmin_slsqp` optimizer returning wrong `imode` ouput.
-        # 1.8.0: `scipy.spatial.qhull._Qhull` is no longer exposed.
-        "scipy>=1.2.0,<1.8.0",
-        # Standard library to generate figures.
-        "matplotlib",
-        # Used internally to read HDF5 format log files.
-        # No wheel is distributed for PyPy on pypi, but pip is able to build
-        # from source without additionnal dependencies.
-        "h5py",
-        # Used internally by Robot to replace meshes by associated minimal
-        # volume bounding box.
-        # No wheel is distributed for PyPy on pypi, and pip is only able to
-        # build from source after install `hdf5-dev` system depdency.
-        "trimesh",
         # Parser for Jiminy's hardware description file.
         "toml",
-        # Web-based mesh visualizer used as Viewer's backend.
-        # 0.0.18 introduces many new features, including loading generic
-        # geometries and jiminy_py viewer relies on it to render collision
-        # bodies.
-        # 0.3.1 updates threejs from 122 to 132, breakin compatibility with
-        # the old, now deprecated, geometry class used to internally to display
-        # tile floor.
-        # 0.3.2 fixes the rendering of DAE meshes.
-        "meshcat>=0.3.2",
+        # Used internally by Robot to replace meshes by associated minimal
+        # volume bounding box.
+        "trimesh",
+        # Used internally to read HDF5 format log files.
+        "h5py",
+        # Use to operate conveniently on nested log data.
+        "dm-tree",
+        # Used internally by Viewer to perform 1D polynomial interpolations.
+        "scipy",
         # Standalone cross-platform mesh visualizer used as Viewer's backend.
         # 1.10.9 adds support of Nvidia EGL rendering without X11 server.
         # Panda3d is NOT supported by PyPy and cannot be built from source.
@@ -137,57 +116,74 @@ setup(
         # Photo-realistic shader for Panda3d to improve rendering of meshes.
         "panda3d_simplepbr",
         # Used internally by Viewer to record video programmatically when
-        # Meshcat is not used as rendering backend.
-        # Cross-platform precompiled binary wheels are provided since 8.0.0.
-        "av",
-        # Used internally by Viewer to detect running Meshcat servers and avoid
-        # orphan child processes.
-        "psutil",
-        # Used internally by Viewer to enable recording video programmatically
-        # while using Meshcat as backend.
-        "pyppeteer",
-        # Used internally by Viewer to send/receive Javascript requests while
-        # recording video using Meshcat backend.
-        # `HTMLSession` is available since 0.3.4.
-        "requests_html>=0.3.4"
+        # Panda3d is used as rendering backend.
+        # >= 8.0.0 provides cross-platform precompiled binary wheels.
+        "av"
     ],
     extras_require={
-      "dev": [
-          # Stub for static type checking
-          "types-toml",
-          # Check PEP8 conformance of Python native code
-          "flake8",
-          # Python linter dependency (Pinned to avoid segfault)
-          "astroid==2.11.2",
-          # Python linter (Pinned to avoid segfault)
-          "pylint==2.13.4",
-          # Python static type checker
-          "mypy>=0.931",
-          # Fix dependency issue with 'sphinx'
-          "jinja2>=3.0,<3.1",
-          # Dependency for documentation generation
-          "pygments",
-          # Dependency for documentation generation
-          "colorama",
-          # Generate Python docs and render '.rst' nicely
-          "sphinx",
-          # 'Read the Docs' Sphinx docs style
-          "sphinx_rtd_theme",
-          # Render markdown in sphinx docs
-          "myst_parser",
-          # Render Jupyter Notebooks in sphinx docs
-          # v0.8.8 introduces a bug for empty 'raw' directives
-          "nbsphinx!=0.8.8",
-          # Render ASCII art diagram (https://aafigure.readthedocs.io)
-          "aafigure",
-          # Bridge between doxygen and sphinx. Used to generate C++ API docs
-          "breathe",
-          # Repair wheels to embed shared libraries.
-          # - 3.2.0: enable defining custom patcher
-          # - 3.3.0: Support Python 3.9 and manylinux_2_24 images
-          # - 4.0.0: Many bug fixes, including RPATH of dependencies
-          "auditwheel>=4.0.0"
-      ]
+        "plot": [
+            # Standard library to generate figures.
+            "matplotlib"
+        ],
+        "meshcat": [
+            # Web-based mesh visualizer used as Viewer's backend.
+            # 0.0.18 introduces many new features, including loading generic
+            # geometries and jiminy_py viewer relies on it to render collision
+            # bodies.
+            # 0.3.1 updates threejs from 122 to 132, breaking compatibility
+            # with the old, now deprecated, geometry class used to internally
+            # to display tile floor.
+            # 0.3.2 fixes the rendering of DAE meshes.
+            "meshcat>=0.3.2",
+            # Used internally by Viewer to read/write Meshcat snapshots.
+            "pillow",
+            # Used internally by Viewer to detect running Meshcat servers and
+            # avoid orphan child processes.
+            "psutil",
+            # Used internally by Viewer to enable recording video
+            # programmatically with Meshcat as backend.
+            "pyppeteer",
+            # Used internally by Viewer to send/receive Javascript requests for
+            # recording video using Meshcat backend.
+            # `HTMLSession` is available since 0.3.4.
+            "requests_html>=0.3.4"
+        ],
+        "dev": [
+            # Stub for static type checking
+            "types-toml",
+            # Check PEP8 conformance of Python native code
+            "flake8",
+            # Python linter dependency (Pinned to avoid segfault)
+            "astroid==2.11.2",
+            # Python linter (Pinned to avoid segfault)
+            "pylint==2.13.4",
+            # Python static type checker
+            "mypy>=0.931",
+            # Fix dependency issue with 'sphinx'
+            "jinja2>=3.0,<3.1",
+            # Dependency for documentation generation
+            "pygments",
+            # Dependency for documentation generation
+            "colorama",
+            # Generate Python docs and render '.rst' nicely
+            "sphinx",
+            # 'Read the Docs' Sphinx docs style
+            "sphinx_rtd_theme",
+            # Render markdown in sphinx docs
+            "myst_parser",
+            # Render Jupyter Notebooks in sphinx docs
+            # v0.8.8 introduces a bug for empty 'raw' directives
+            "nbsphinx!=0.8.8",
+            # Render ASCII art diagram (https://aafigure.readthedocs.io)
+            "aafigure",
+            # Bridge between doxygen and sphinx. Used to generate C++ API docs
+            "breathe",
+            # Repair wheels to embed shared libraries.
+            # - 3.2.0: enable defining custom patcher
+            # - 3.3.0: Support Python 3.9 and manylinux_2_24 images
+            # - 4.0.0: Many bug fixes, including RPATH of dependencies
+            "auditwheel>=4.0.0"
+        ]
     },
     zip_safe=False
 )
