+++ python/jiminy_py/src/jiminy_py/viewer/__init__.py
@@ -1,6 +1,5 @@
 from .viewer import Viewer, sleep
-from .replay import (TrajectoryDataType,
-                     extract_replay_data_from_log_data,
+from .replay import (extract_replay_data_from_log_data,
                      play_trajectories,
                      play_logs_data,
                      play_logs_files)
@@ -8,7 +7,6 @@ from .meshcat.utilities import interactive_mode
 
 
 __all__ = [
-    'TrajectoryDataType',
     'sleep',
     'Viewer',
     'interactive_mode',
